requirements.txt
apache-airflow==2.9.0
pandas
scikit-learn
joblib
requests
python-dotenv
google-cloud-storage
